feat: implement house-edge crash distribution

+++ backend/src/main/java/ru/hackathon/airballoon/config/GameConfig.java
@@ -7,7 +7,7 @@ import java.util.List;
 public record GameConfig(
     String gameId, String gameName, String gameType, boolean active,
     int greenLevelCount, int redLevelCount,
-    BigDecimal minCrashMultiplier, BigDecimal maxCrashMultiplier, double growthRate, double alpha,
+    BigDecimal minCrashMultiplier, BigDecimal maxCrashMultiplier, double growthRate, BigDecimal alpha,
     int updateIntervalMs, BigDecimal minBet, BigDecimal maxBet, List<Integer> boosterValues,
     List<Integer> greenBoosterWeights, List<Integer> redBoosterWeights,
     long pointsPerLevel, long pointsCashoutBonus, long pointsX2Bonus, long pointsX3Bonus, long pointsX4Bonus,
